Delete conflicting files to fix jazzy docset generation (#2076)
* Deleted conflicting Date.swift

There were two Date.swift files in the merged, CocoaPods-generated target that jazzy uses to calculate USRs.

* Resolved more file name collisions
