--- MapboxNavigation/Date.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-extension Date {
-    
-    static let ISO8601Formatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        return formatter
-    }()
-}
